Moves xpipe/straw to common from redalert.
